feature: responsibility controller

+++ src/main/java/pro/sorokovsky/schoolmanagerbackend/service/RequirementsService.java
@@ -47,10 +47,10 @@ public class RequirementsService {
 
     @Transactional
     public void deleteById(Integer id) {
-        try {
+        if (repository.existsById(id)) {
             repository.deleteById(id);
-        } catch (Exception e) {
-            e.printStackTrace();
+        } else {
+            throw new RequirementNotFoundException();
         }
     }
 }
