+++ src/main/java/pro/sorokovsky/schoolmanagerbackend/service/ResponsibilitiesService.java
@@ -46,6 +46,10 @@ public class ResponsibilitiesService {
     }
 
     public void delete(Integer id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new ResponsibilityNotFoundException();
+        }
     }
 }
